Rework of location_ui_serverside.lua
Now supports multiple location tables for different maps and will initiate the table depending on what map you are on. This is useful if you use multiple maps and don't want to have to change every location (in shared) every time you change map. The name of the lua file in shared must be "mapnamelocations.lua". For example, "rp_city9locations.lua".